`Direction` moved to directive root namespace.
