+++ packages/graphql/src/SortBy/Sorters/EloquentSorter.php
@@ -19,6 +19,7 @@ use Illuminate\Database\Query\JoinClause;
 use LastDragon_ru\LaraASP\Core\Utils\Cast;
 use LastDragon_ru\LaraASP\Eloquent\ModelHelper;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Property;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Direction;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Exceptions\RelationUnsupported;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Nulls;
 use LogicException;
