+++ packages/graphql/src/SortBy/Types/Direction.php
@@ -8,8 +8,8 @@ use GraphQL\Type\Definition\Type;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\TypeDefinition;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\TypeSource;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Manipulator;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Direction as DirectionEnum;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Directives\Directive;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\Direction as DirectionEnum;
 use Override;
 
 class Direction implements TypeDefinition {
